+++ demo/sample.js
@@ -297,6 +297,7 @@ $(function(){
 				  {name: "Vista (classic Dynatree)", value: "vista", href: "skin-vista/ui.fancytree.css"},
 				  {name: "Win7", value: "win7", href: "skin-win7/ui.fancytree.css"},
 				  {name: "Win8", value: "win8", href: "skin-win8/ui.fancytree.css"},
+				  {name: "Win8-N", value: "win8n", href: "skin-win8-n/ui.fancytree.css"},
 				  {name: "Win8 xxl", value: "win8xxl", href: "skin-win8-xxl/ui.fancytree.css"},
 				  {name: "Lion", value: "lion", href: "skin-lion/ui.fancytree.css"}
 				  ]